server/requirements.txt
flask
python-dotenv
sqlalchemy
sqlalchemy-utils
psycopg2
flask-cors
APScheduler